# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing passengers and taxists in a ride-sharing application.",
        "title": "Ride-Sharing Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.55.42:8080",
    "basePath": "/",
    "paths": {
...
        "/protected/taxist-announcements/{departed}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all taxist announcements by given departed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get all specific taxist announcements by departed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Departed",
                        "name": "departed",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/protected/taxist-comments/{taxist_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all taxist comments by given taxist ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get all specific taxist comments by taxist_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Taxist ID",
                        "name": "taxist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "add a new comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "POST a new comment for passenger user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Taxist ID",
                        "name": "taxist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Passenger Comment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostComment"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/protected/taxist-departed/{taxi_ann_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a taxist's announcement department by taxi_ann_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Update taxist announcement department",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Announcement ID",
                        "name": "taxi_ann_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/protected/taxist-notifications/{taxist_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all taxist notifications by given taxist_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Announcement"
                ],
                "summary": "Get all specific taxist notifications by taxist_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Taxist ID",
                        "name": "taxist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "handlers.PostComment": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}